Many changes
Primarily in regards to the Video Player, which functions correctly, breaking MVVM design pattern rules as little as possible. Some other miscellaneous changes I thought were necessary, such as the removal of Windows Visibility property exposure, with boolean type conversion replacing it. 6.0.0-alpha of YoutubeExplode was released, so added the new NuGet package for automatic updates and compatibility with YoutubeExplode.Converter. All-in-all, a productive commit.